avoid some other potential #events overflows
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@15739 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ framework/modules/progressbar/include/ProgressBarModule.h
@@ -31,9 +31,9 @@ namespace Belle2 {
 
   protected:
     /** Number of processed events. */
-    int m_evtNr;
+    long m_evtNr;
     /** Total number of events. */
-    int m_nTotal;
+    long m_nTotal;
     /** Start time (ns). */
     long m_startTime;
     /** Timestamp when we last printed something (ns). */
